dereferencing a pointer at 219
This will create a new SDL_Rect variable called speed_rect, and initialize all of its fields to zero. The SDL_Rect struct has four fields, which represent the x and y coordinates of the rectangle's top-left corner, and the width and height of the rectangle, respectively. By setting all of these fields to zero, we effectively create an SDL_Rect that represents a rectangle with a width and height of zero, positioned at the origin (0,0).